refactor: Separate CurrentPost from post.js
Post.js에 있는 CurrentPost를 다른 파일로 분리했습니다.
페이지네이션 라이브러리를 시도중입니다.
아직 전과 동일하기 때문에 그대로 머지합니다.
