+++ frontend/package.json
@@ -14,12 +14,13 @@
     "axios": "^1.3.3",
     "date-fns": "^2.30.0",
     "eslint": "8.22.0",
-    "moment": "^2.29.4",
     "history": "^5.3.0",
+    "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-datepicker": "^4.14.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.1",
+    "react-js-pagination": "^3.0.3",
     "react-quill": "^2.0.0",
     "react-redux": "^8.0.5",
     "react-router": "^6.8.1",
